Clean up connection wrapper code
This creates a .hh file to put all of the functionality of connection
in. I also made the code DRY, and now closed all the fabric objects at
the end. Finally I fixed the bug causing the message receiver to not get
the message. connection.cc now contains a basic example of connecting
and sending a message (for testing purposes)