Create tests for connection.hh
Previously wasn't tests connection.hh which is bad. Now there are tests,
which are loosely based on connectionless.hh. They all pass. I had to
add another constructor to connection.hh for purposes of the tests, so
that the client knows when the server is listening for a connection
(without this it may attempt to connect to early and crash)