Update connection.hh try_ methods
Previously if there was an error reported on the cq in any of the
try_ method calls, the SAFE_CALL around wait_for_completion would
cause the program to exit. This was bad as many of the errors were
reported on the completion queue. This commit updates it so that
if an was found on the cq, the try_ methods return false instead
of exiting.